requirements.txt
absl-py==1.4.0
#albumentations==0.1.12
arabic-reshaper==2.1.3
arrow==1.2.3
attrs==22.2.0
audioread==3.0.0
Augmentor==0.2.12
augmixations==0.1.2
Babel==2.12.1
beautifulsoup4==4.12.2
bezier==2021.2.12
boto3==1.26.114
botocore==1.29.114
bravado==11.0.3
bravado-core==5.17.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
clldutils==3.19.0
colorama==0.4.6
colorlog==6.7.0
csvw==3.1.3
cycler==0.11.0
cytoolz==0.12.1
dataclasses==0.6
decorator==5.1.1
diffimg==0.2.3
dill==0.3.6
dlinfo==1.2.1
editdistance==0.6.2
einops==0.6.0
encodec==0.1.1
filelock==3.11.0
fonttools==4.39.3
fqdn==1.5.1
future==0.18.2
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.2
google-auth-oauthlib==1.0.0
GPUtil==1.4.0
graphviz==0.20.1
grpcio==1.53.0
huggingface-hub==0.13.4
hwb==0.0.15
idna==3.4
imageio==2.9.0
imgaug==0.2.6
importlib-metadata==6.3.0
importlib-resources==5.12.0
intervaltree==3.1.0
isodate==0.6.1
isoduration==20.11.0
jmespath==1.0.1
joblib==1.2.0
jsonpointer==2.3
jsonref==1.1.0
jsonschema==4.17.3
#k2 @ https://huggingface.co/csukuangfj/k2/resolve/main/cuda/k2-1.23.4.dev20230224+cuda11.6.torch1.13.0-cp38-cp38-linux_x86_64.whl
#kaldialign==0.6
#kaldifst==1.6
#kaldilm==1.15
kiwisolver==1.4.4
language-tags==1.2.0
#lhotse @ git+https://github.com/lhotse-speech/lhotse@8859200bdfc8bad8c528be8598ce42ab3334d28e
librosa==0.8.1
lilcom==1.6
llvmlite==0.39.1
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.5.3
monotonic==1.6
msgpack==1.0.5
neptune-client==0.16.15
networkx==3.1
#numba==0.56.4
numpy==1.20.3
oauthlib==3.2.2
opencv-python==4.4.0.46
packaging==23.0
pandas==1.1.4
phonemizer==3.2.1
Pillow==9.3.0
pkg_resources==0.0.0
pkgutil_resolve_name==1.3.10
platformdirs==3.2.0
pooch==1.7.0
progressbar2==3.53.1
protobuf==3.20.1
psutil==6.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pylatexenc==2.10
pyparsing==3.0.9
pyrsistent==0.19.3
python-bidi==0.4.2
python-dateutil==2.8.1
python-utils==3.5.2
pytz==2020.4
PyWavelets==1.4.1
PyYAML==6.0
rdflib==6.3.2
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
resampy==0.4.2
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3987==1.3.8
rsa==4.9
s3transfer==0.6.0
scikit-image==0.19.3
scikit-learn==1.2.2
scipy==1.10.1
segments==2.2.1
sentencepiece==0.1.97
simplejson==3.19.1
six==1.15.0
smmap==5.0.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4.1
swagger-spec-validator==3.0.3
tabulate==0.9.0
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
threadpoolctl==3.1.0
tifffile==2023.4.12
tokenizers==0.13.2
toolz==0.12.0
#torch==1.13.0+cu116
#torchaudio==0.13.0+cu116
#torchmetrics==0.11.0
torchtext==0.8.0
#torchvision==0.14.0+cu116
tqdm==4.65.0
transformers==4.25.1
trdg==1.8.0
typeguard==3.0.2
typing_extensions==4.4.0
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.15
#-e git+https://github.com/lifeiteng/valle.git@6d1e8e2c61b6ecee32552f96bc12a303eaef686c#egg=valle
webcolors==1.13
websocket-client==1.5.1
Werkzeug==2.2.3
wikipedia==1.4.0
zipp==3.15.0
